requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pymupdf
Pillow
pytesseract
numpy
pydantic
tqdm